+++ src/messages/ru/admin.php
@@ -377,7 +377,7 @@ return [
     'active_button_duplicate_success' => 'Копия была создана.',
     'active_button_duplicate_error' => 'Ошибка при создании копии: {message}',
 
-// 1.3
+// 2.0
 
     'model_tag_translation' => 'Translation',
     'model_tag_translation_hint' => 'An optional translation for the tag in the given language. This allows to assign and output the same tag on multi-language websites.',
@@ -393,4 +393,6 @@ return [
     'js_scheduler_table_newvalue' => 'New value',
     'js_scheduler_table_timestamp' => 'Execution Time',
     'js_dir_manager_rename_success' => 'The folder has been renamed successfully.',
+    'layout_filemanager_search_no_results' => 'No files were found for <strong>{{searchQuery}}</strong>.',
+    'layout_filemanager_empty_folder' => 'This directory contains no files.',
 ];